Test macOS on canary with separate workflow (#15901)
Follow up to #15899 since it didn't solve the [concurrency limit](https://docs.github.com/en/actions/getting-started-with-github-actions/about-github-actions#usage-limits) issue.

It seems that the job is queued prior to the `if` statement being evaluated so we must use a separate workflow to listen for push events and filter prior to the job being queued.